moved and improved send sync tests
