+++ identity_resolver/src/resolution/tests/mod.rs
@@ -1,0 +1,7 @@
+// Copyright 2020-2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use super::resolver::*;
+mod failures;
+mod send_sync;
+mod success;
