+++ identity_resolver/src/resolution/tests/success.rs
@@ -1,0 +1,2 @@
+// Copyright 2020-2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
